[New feature] New vip status depend on backends status

	modified:   usr/share/perl5/Zevenet/API3/Farm/Get.pm
	modified:   usr/share/perl5/Zevenet/API3/Farm/Get/Datalink.pm
	modified:   usr/share/perl5/Zevenet/API3/Farm/Get/GSLB.pm
	modified:   usr/share/perl5/Zevenet/API3/Farm/Get/HTTP.pm
	modified:   usr/share/perl5/Zevenet/API3/Farm/Get/L4xNAT.pm
	modified:   usr/share/perl5/Zevenet/API3/Stats.pm
	modified:   usr/share/perl5/Zevenet/Farm/Base.pm
	modified:   usr/share/perl5/Zevenet/Farm/HTTP/Action.pm
	modified:   usr/share/perl5/Zevenet/Farm/HTTP/Backend.pm

+++ usr/share/perl5/Zevenet/Farm/HTTP/Action.pm
@@ -99,9 +99,16 @@ sub _runHTTPFarmStop    # ($farm_name)
 		my $pid = &getFarmPid( $farm_name );
 		my $piddir = &getGlobalConfiguration('piddir');
 
-		&zenlog( "running 'kill 15, $pid'" );
-		my $run = kill 15, $pid;
-		$status = $?;
+		if ( $pid eq '-' || $pid == -1 )
+		{
+			&zenlog( "Not found pid" );
+		}
+		else
+		{
+			&zenlog( "running 'kill 15, $pid'" );
+			my $run = kill 15, $pid;
+			$status = $?;
+		}
 
 		unlink ( "$piddir\/$farm_name\_pound.pid" ) if -e "$piddir\/$farm_name\_pound.pid";
 		unlink ( "\/tmp\/$farm_name\_pound.socket" ) if -e "\/tmp\/$farm_name\_pound.socket";
